refactored footer actions to accommodate more buttons
