+++ packages/kokoas-client/src/pages/projEstimate/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from './useConfirmBeforeClose';
 export * from './useFormReset';
 export * from './useResolveParam';
 export * from './useRowValues';
